requirements.txt
flask
flask-cors
requests
python-dotenv
openai
groq
langchain
langgraph
pandas
numpy
